added another helper function to make creating raw multisig transactions easier, also various small fixes

+++ src/bitcoin-api/ibtcrpc.hpp
@@ -45,9 +45,9 @@ struct BitcoinServer
     std::string user;
     std::string password;
     std::string url;
-    uint64_t port;
+    int32_t port;
 
-    BitcoinServer(const std::string &account, const std::string &password, const std::string &url, const uint64_t &port):
+    BitcoinServer(const std::string &account, const std::string &password, const std::string &url, const uint32_t &port):
         user(account), password(password), url(url), port(port)
     {}
 };
